+++ database/migrations/2023_02_03_100714_create_applicant_guardians_table.php
@@ -21,7 +21,7 @@ return new class extends Migration
             $table->string('guardian_lname');
             $table->string('guardian_religion');
             $table->string('guardian_occupation');
-            $table->unsignedInteger('guardian_annual_income');
+            $table->float('guardian_annual_income');
             $table->string('guardian_phone_num');
             $table->timestamps();
         });
